requirements.txt
streamlit>=1.28.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
chromadb>=0.4.0
python-dateutil>=2.8.2
dateparser>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0